added stlye images, selector, and indicator.
