+++ src/components/productDetailAttempt/productStyle.css
@@ -5,7 +5,8 @@
   border-radius:1%;
   border: solid .25em #97d1cf;
   flex-wrap: wrap;
-  margin-top: 1vh;
+  margin-top: .5vh;
+  margin-bottom: .5vh;
   cursor: pointer;
 
 }
@@ -30,6 +31,22 @@
   /* Firefox */
 }
 
+.styleThumbnail {
+  width: min(75px, 5vw);
+  height: min(75px, 5vw);
+  border-radius:100%;
+  border: solid .25em #97d1cf;
+  flex-wrap: wrap;
+  margin-top: .5vh;
+  margin-bottom: .5vh;
+  cursor: pointer;
+  position: relative;
+}
+
+.styleThumbnailWrapper {
+  width: 250px;
+}
+
 .mainImg {
   height: 100%;
   width: 100%;
@@ -39,7 +56,7 @@
 .mainImgWrapper {
   position: relative;
   display:flex;
-  min-width:50vw;
+  width: min(700px, 50vw);
 }
 
 .galleryWrapper {
@@ -57,8 +74,29 @@ display: flex;
 .detailWrapper {
   width:100%;
   min-width: 20vw;
+  padding-left: 1vw;
 }
 
 .infoWrapper {
   position: relative;
+  height: 30vh;
 }
+
+.infoDetailCategory {
+  position: relative;
+  color:teal;
+  font-size: 1.25rem;
+  margin-top: 2vh;
+}
+.infoDetailName {
+  position: relative;
+  color:#2088ff;
+  font-size: 2rem;
+  margin-top: 2vh;
+}
+.infoDetailPrice {
+  margin-top: 2vh;
+}
+
+
+
